backend/requirements.txt
fastapi
opencv-python
uvicorn[standard]
motor
pydantic[email]
python-dotenv
passlib[bcrypt]
Authlib